requirements.txt

flask==3.0.3
werkzeug==3.0.4

# Whisper (local STT)
openai-whisper==20231117
numpy>=1.23.0
torch>=2.0.0

# Gemini
google-generativeai==0.7.2

# Google Docs API
google-api-python-client==2.137.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Audio processing helper
pydub==0.25.1

# Env
python-dotenv==1.0.1
